Use struct Material to bundle material properties
Material properties of a IObject3D are bundled inside a struct Material.
These include (for now) color and emission and will allow encoding the type
of surface (diffusive, reflective or refractive). Since the material is a
property of the IObject3D the definition of struct Material has been moved
to Objects.hpp