Refactor render method
After refactoring IRenderer it makes sense to have the Surface as
a member of the Camera object. The IRenderer user must provide a
reference to a Camera object to render the scene so instead of
having renderScene() return a pointer to a new Surface, we use
a reference to the pre-allocated Surface in the Camera.