Add IResultsListener interface
The PathTracer can now notify listeners of partial and final rendering
results. This allows, for instance, to update a window when a part of
the surface is complete.

Using this IResultsInterface, rendering is now handled in blocks and
windows can display the rendering progress as soon as it receives
a result notification.